requirements.txt
python-telegram-bot==20.3
IMDbPY==2022.1.6
requests==2.26.0
